[FAG] modify tilingdata dtype

+++ attention/flash_attention_score_grad/op_kernel/arch35/flash_attention_score_grad_tiling_data_regbase.h
@@ -278,24 +278,24 @@ class PreParamsRegbase {
 public:
     uint64_t maskPreBlockTotal;
     uint64_t maskSingleCoreNum;
-    uint32_t qPreBlockFactor;
+    uint64_t qPreBlockFactor;
     uint64_t qPreBlockTotal;
-    uint32_t qPreBlockTail;
-    uint32_t kPreBlockFactor;
+    uint64_t qPreBlockTail;
+    uint64_t kPreBlockFactor;
     uint64_t kPreBlockTotal;
-    uint32_t kPreBlockTail;
-    uint32_t vPreBlockFactor;
+    uint64_t kPreBlockTail;
+    uint64_t vPreBlockFactor;
     uint64_t vPreBlockTotal;
-    uint32_t vPreBlockTail;
-    uint32_t maskCoreNum;
-    uint32_t castBufferLen;
-    uint32_t outputBufferLen;
-    uint32_t inputBufferLen;
-    uint32_t singleUBProcessNum;
-    uint32_t maskSingleCoreLoop;
-    uint32_t maskLastLoopNum;
-    uint32_t maskTailCoreLoop;
-    uint32_t maskTailCoreLastLoopNum;
+    uint64_t vPreBlockTail;
+    uint64_t maskCoreNum;
+    uint64_t castBufferLen;
+    uint64_t outputBufferLen;
+    uint64_t inputBufferLen;
+    uint64_t singleUBProcessNum;
+    uint64_t maskSingleCoreLoop;
+    uint64_t maskLastLoopNum;
+    uint64_t maskTailCoreLoop;
+    uint64_t maskTailCoreLastLoopNum;
     uint32_t dropoutIsDivisibleBy8;
     bool sValueZeroUnderTND;
     uint8_t reserved1; // tilingData需要8字节对齐
@@ -318,24 +318,24 @@ public:
 
     uint64_t get_maskPreBlockTotal() const { return maskPreBlockTotal; }
     uint64_t get_maskSingleCoreNum() const { return maskSingleCoreNum; }
-    uint32_t get_qPreBlockFactor() const { return qPreBlockFactor; }
+    uint64_t get_qPreBlockFactor() const { return qPreBlockFactor; }
     uint64_t get_qPreBlockTotal() const { return qPreBlockTotal; }
-    uint32_t get_qPreBlockTail() const { return qPreBlockTail; }
-    uint32_t get_kPreBlockFactor() const { return kPreBlockFactor; }
+    uint64_t get_qPreBlockTail() const { return qPreBlockTail; }
+    uint64_t get_kPreBlockFactor() const { return kPreBlockFactor; }
     uint64_t get_kPreBlockTotal() const { return kPreBlockTotal; }
-    uint32_t get_kPreBlockTail() const { return kPreBlockTail; }
-    uint32_t get_vPreBlockFactor() const { return vPreBlockFactor; }
+    uint64_t get_kPreBlockTail() const { return kPreBlockTail; }
+    uint64_t get_vPreBlockFactor() const { return vPreBlockFactor; }
     uint64_t get_vPreBlockTotal() const { return vPreBlockTotal; }
-    uint32_t get_vPreBlockTail() const { return vPreBlockTail; }
-    uint32_t get_maskCoreNum() const { return maskCoreNum; }
-    uint32_t get_castBufferLen() const { return castBufferLen; }
-    uint32_t get_outputBufferLen() const { return outputBufferLen; }
-    uint32_t get_inputBufferLen() const { return inputBufferLen; }
-    uint32_t get_singleUBProcessNum() const { return singleUBProcessNum; }
-    uint32_t get_maskSingleCoreLoop() const { return maskSingleCoreLoop; }
-    uint32_t get_maskLastLoopNum() const { return maskLastLoopNum; }
-    uint32_t get_maskTailCoreLoop() const { return maskTailCoreLoop; }
-    uint32_t get_maskTailCoreLastLoopNum() const { return maskTailCoreLastLoopNum; }
+    uint64_t get_vPreBlockTail() const { return vPreBlockTail; }
+    uint64_t get_maskCoreNum() const { return maskCoreNum; }
+    uint64_t get_castBufferLen() const { return castBufferLen; }
+    uint64_t get_outputBufferLen() const { return outputBufferLen; }
+    uint64_t get_inputBufferLen() const { return inputBufferLen; }
+    uint64_t get_singleUBProcessNum() const { return singleUBProcessNum; }
+    uint64_t get_maskSingleCoreLoop() const { return maskSingleCoreLoop; }
+    uint64_t get_maskLastLoopNum() const { return maskLastLoopNum; }
+    uint64_t get_maskTailCoreLoop() const { return maskTailCoreLoop; }
+    uint64_t get_maskTailCoreLastLoopNum() const { return maskTailCoreLastLoopNum; }
     uint32_t get_dropoutIsDivisibleBy8() const { return dropoutIsDivisibleBy8; }
     bool get_sValueZeroUnderTND() const { return sValueZeroUnderTND; }
     uint32_t get_sfmgUsedCoreNum() const {return sfmgUsedCoreNum;}
@@ -353,24 +353,24 @@ public:
 
     void set_maskPreBlockTotal(uint64_t val) { maskPreBlockTotal = val; }
     void set_maskSingleCoreNum(uint64_t val) { maskSingleCoreNum = val; }
-    void set_qPreBlockFactor(uint32_t val) { qPreBlockFactor = val; }
+    void set_qPreBlockFactor(uint64_t val) { qPreBlockFactor = val; }
     void set_qPreBlockTotal(uint64_t val) { qPreBlockTotal = val; }
-    void set_qPreBlockTail(uint32_t val) { qPreBlockTail = val; }
-    void set_kPreBlockFactor(uint32_t val) { kPreBlockFactor = val; }
+    void set_qPreBlockTail(uint64_t val) { qPreBlockTail = val; }
+    void set_kPreBlockFactor(uint64_t val) { kPreBlockFactor = val; }
     void set_kPreBlockTotal(uint64_t val) { kPreBlockTotal = val; }
-    void set_kPreBlockTail(uint32_t val) { kPreBlockTail = val; }
-    void set_vPreBlockFactor(uint32_t val) { vPreBlockFactor = val; }
+    void set_kPreBlockTail(uint64_t val) { kPreBlockTail = val; }
+    void set_vPreBlockFactor(uint64_t val) { vPreBlockFactor = val; }
     void set_vPreBlockTotal(uint64_t val) { vPreBlockTotal = val; }
-    void set_vPreBlockTail(uint32_t val) { vPreBlockTail = val; }
-    void set_maskCoreNum(uint32_t val) { maskCoreNum = val; }
-    void set_castBufferLen(uint32_t val) { castBufferLen = val; }
-    void set_outputBufferLen(uint32_t val) { outputBufferLen = val; }
-    void set_inputBufferLen(uint32_t val) { inputBufferLen = val; }
-    void set_singleUBProcessNum(uint32_t val) { singleUBProcessNum = val; }
-    void set_maskSingleCoreLoop(uint32_t val) { maskSingleCoreLoop = val; }
-    void set_maskLastLoopNum(uint32_t val) { maskLastLoopNum = val; }
-    void set_maskTailCoreLoop(uint32_t val) { maskTailCoreLoop = val; }
-    void set_maskTailCoreLastLoopNum(uint32_t val) { maskTailCoreLastLoopNum = val; }
+    void set_vPreBlockTail(uint64_t val) { vPreBlockTail = val; }
+    void set_maskCoreNum(uint64_t val) { maskCoreNum = val; }
+    void set_castBufferLen(uint64_t val) { castBufferLen = val; }
+    void set_outputBufferLen(uint64_t val) { outputBufferLen = val; }
+    void set_inputBufferLen(uint64_t val) { inputBufferLen = val; }
+    void set_singleUBProcessNum(uint64_t val) { singleUBProcessNum = val; }
+    void set_maskSingleCoreLoop(uint64_t val) { maskSingleCoreLoop = val; }
+    void set_maskLastLoopNum(uint64_t val) { maskLastLoopNum = val; }
+    void set_maskTailCoreLoop(uint64_t val) { maskTailCoreLoop = val; }
+    void set_maskTailCoreLastLoopNum(uint64_t val) { maskTailCoreLastLoopNum = val; }
     void set_dropoutIsDivisibleBy8(uint32_t val) { dropoutIsDivisibleBy8 = val; }
     void set_sValueZeroUnderTND(bool val) { sValueZeroUnderTND = val; }
     void set_sfmgUsedCoreNum(uint32_t val) { sfmgUsedCoreNum = val; }
@@ -389,19 +389,19 @@ public:
 
 class PostParamsRegbase {
 public:
-    uint32_t postUbBaseSize;
-    uint32_t qPostBlockFactor;
+    uint64_t postUbBaseSize;
+    uint64_t qPostBlockFactor;
     uint64_t qPostBlockTotal;
-    uint32_t qPostBaseNum;
-    uint32_t qPostTailNum;
-    uint32_t kPostBlockFactor;
+    uint64_t qPostBaseNum;
+    uint64_t qPostTailNum;
+    uint64_t kPostBlockFactor;
     uint64_t kPostBlockTotal;
-    uint32_t kPostBaseNum;
-    uint32_t kPostTailNum;
-    uint32_t vPostBlockFactor;
+    uint64_t kPostBaseNum;
+    uint64_t kPostTailNum;
+    uint64_t vPostBlockFactor;
     uint64_t vPostBlockTotal;
-    uint32_t vPostBaseNum;
-    uint32_t vPostTailNum;
+    uint64_t vPostBaseNum;
+    uint64_t vPostTailNum;
     uint64_t dqWorkSpaceOffset;
     uint64_t dkWorkSpaceOffset;
     uint64_t dvWorkSpaceOffset;
@@ -411,19 +411,19 @@ public:
     uint64_t deterWorkSpaceOffset;
     uint64_t sfmgWorkSpaceOffset;
 
-    uint32_t get_postUbBaseSize() const { return postUbBaseSize; }
-    uint32_t get_qPostBlockFactor() const { return qPostBlockFactor; }
+    uint64_t get_postUbBaseSize() const { return postUbBaseSize; }
+    uint64_t get_qPostBlockFactor() const { return qPostBlockFactor; }
     uint64_t get_qPostBlockTotal() const { return qPostBlockTotal; }
-    uint32_t get_qPostBaseNum() const { return qPostBaseNum; }
-    uint32_t get_qPostTailNum() const { return qPostTailNum; }
-    uint32_t get_kPostBlockFactor() const { return kPostBlockFactor; }
+    uint64_t get_qPostBaseNum() const { return qPostBaseNum; }
+    uint64_t get_qPostTailNum() const { return qPostTailNum; }
+    uint64_t get_kPostBlockFactor() const { return kPostBlockFactor; }
     uint64_t get_kPostBlockTotal() const { return kPostBlockTotal; }
-    uint32_t get_kPostBaseNum() const { return kPostBaseNum; }
-    uint32_t get_kPostTailNum() const { return kPostTailNum; }
-    uint32_t get_vPostBlockFactor() const { return vPostBlockFactor; }
+    uint64_t get_kPostBaseNum() const { return kPostBaseNum; }
+    uint64_t get_kPostTailNum() const { return kPostTailNum; }
+    uint64_t get_vPostBlockFactor() const { return vPostBlockFactor; }
     uint64_t get_vPostBlockTotal() const { return vPostBlockTotal; }
-    uint32_t get_vPostBaseNum() const { return vPostBaseNum; }
-    uint32_t get_vPostTailNum() const { return vPostTailNum; }
+    uint64_t get_vPostBaseNum() const { return vPostBaseNum; }
+    uint64_t get_vPostTailNum() const { return vPostTailNum; }
     uint64_t get_dqWorkSpaceOffset() const { return dqWorkSpaceOffset; }
     uint64_t get_dkWorkSpaceOffset() const { return dkWorkSpaceOffset; }
     uint64_t get_dvWorkSpaceOffset() const { return dvWorkSpaceOffset; }
@@ -433,19 +433,19 @@ public:
     uint64_t get_deterWorkSpaceOffset() const { return deterWorkSpaceOffset; }
     uint64_t get_sfmgWorkSpaceOffset() const {return sfmgWorkSpaceOffset;}
 
-    void set_postUbBaseSize(uint32_t value) { postUbBaseSize = value; }
-    void set_qPostBlockFactor(uint32_t value) { qPostBlockFactor = value; }
+    void set_postUbBaseSize(uint64_t value) { postUbBaseSize = value; }
+    void set_qPostBlockFactor(uint64_t value) { qPostBlockFactor = value; }
     void set_qPostBlockTotal(uint64_t value) { qPostBlockTotal = value; }
-    void set_qPostBaseNum(uint32_t value) { qPostBaseNum = value; }
-    void set_qPostTailNum(uint32_t value) { qPostTailNum = value; }
-    void set_kPostBlockFactor(uint32_t value) { kPostBlockFactor = value; }
+    void set_qPostBaseNum(uint64_t value) { qPostBaseNum = value; }
+    void set_qPostTailNum(uint64_t value) { qPostTailNum = value; }
+    void set_kPostBlockFactor(uint64_t value) { kPostBlockFactor = value; }
     void set_kPostBlockTotal(uint64_t value) { kPostBlockTotal = value; }
-    void set_kPostBaseNum(uint32_t value) { kPostBaseNum = value; }
-    void set_kPostTailNum(uint32_t value) { kPostTailNum = value; }
-    void set_vPostBlockFactor(uint32_t value) { vPostBlockFactor = value; }
+    void set_kPostBaseNum(uint64_t value) { kPostBaseNum = value; }
+    void set_kPostTailNum(uint64_t value) { kPostTailNum = value; }
+    void set_vPostBlockFactor(uint64_t value) { vPostBlockFactor = value; }
     void set_vPostBlockTotal(uint64_t value) { vPostBlockTotal = value; }
-    void set_vPostBaseNum(uint32_t value) { vPostBaseNum = value; }
-    void set_vPostTailNum(uint32_t value) { vPostTailNum = value; }
+    void set_vPostBaseNum(uint64_t value) { vPostBaseNum = value; }
+    void set_vPostTailNum(uint64_t value) { vPostTailNum = value; }
     void set_dqWorkSpaceOffset(uint64_t value) { dqWorkSpaceOffset = value; }
     void set_dkWorkSpaceOffset(uint64_t value) { dkWorkSpaceOffset = value; }
     void set_dvWorkSpaceOffset(uint64_t value) { dvWorkSpaceOffset = value; }
